Update moved Field to its own submodule
