+++ las/__init__.py
@@ -3,7 +3,8 @@ import logging
 from .api_client import ApiClient
 from .client import Client
 from .credentials import Credentials
-from .prediction import Prediction, Field
+from .prediction import Prediction
+from .field import Field
 
 
 logger = logging.getLogger('las')
